requirements.txt
Django==6.0
djangorestframework==3.16.1
phonenumbers==9.0.21
djangorestframework-simplejwt==5.4.0
twilio==9.9.0

# Images (pour ImageField)
Pillow==11.1.0

# PostgreSQL
psycopg2-binary==2.9.10

# Environnement
python-dotenv==1.1.0

# Serveur WSGI (production)
gunicorn==23.0.0

# CORS headers
django-cors-headers==4.7.0
requests==2.32.5
structlog==22.1.0

# Gestion des pays et devises
pycountry==24.6.1